trading-ai-analysis/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
plotly>=5.18.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
schedule>=1.2.0
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
python-dateutil>=2.8.2
requests>=2.31.0
ta>=0.10.0
yfinance>=0.2.0
torch>=2.1.0
transformers>=4.35.0
python-binance>=1.0.0
ccxt>=4.1.0
safetensors>=0.4.0
psutil>=5.9.0
accelerate>=0.25.0